Avoid TypeErrors
When user cancels input (prompt() returns null),
TypeErrors are thrown because of use of
toUpperCase(), a string method. Further when user
submits an empty string, it is irrelevant to the
game, so humanChoice is not returned. Thus
playRound function gets humanChoice as undefined,
which again because of use of toUpperCase()
raises TypeErrors.